requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
Flask-Login>=0.6
python-dotenv>=1.0
Werkzeug>=2.2
gunicorn>=20.0
pymysql>=1.0
